backend/app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.1.0"
description = "Mitten API for environmental analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.1.0",
    "uvicorn>=0.37.0",
    "orjson>=3.9.15",
    "google-generativeai>=0.3.2",
    "earthengine-api>=1.6.9",
    "fastapi>=0.117.1",
    "folium>=0.20.0",
    "leafmap>=0.52.4",
    "plotly>=6.3.0",
    "pydeck>=0.9.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.24.1"  # For testing FastAPI endpoints
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["app"]
